think we have old functionality working again

+++ download.py
@@ -34,6 +34,10 @@ class DownloadClient:
 
     def download(self, movie_obj):
         self.movie = movie_obj
+        self.vvn1_mongo_client.update_downloading_status(self.guild_id,self.movie,True)
+        if not self.vvn1_mongo_client.status_update_id in self.movie:
+            self.movie[self.vvn1_mongo_client.status_update_id] = self.vvn1_mongo_client.create_status_update_obj(self.guild_id,self.movie['_id'],self.vvn1_mongo_client.DOWNLOADING,self.movie['userID'],self.movie['textChannelID'])
+
         torrent_url = movie_obj['torrentLink']
 
         pia_conn = self.check_pia()
@@ -41,9 +45,12 @@ class DownloadClient:
             return 'Not connected to VPN. Quitting...'
         movie_dir_name = self.download_torrent(torrent_url)
 
-        final_dir = self.fix_dir(movie_dir_name)
-
-        self.vvn1_mongo_client.download_successful(self.guild_id, self.movie)
+        if 'path' not in self.movie:
+            final_dir = self.fix_dir(movie_dir_name)
+        else:
+            final_dir = self.movie['path']
+        final_dir_formatted = self.zip_movie(final_dir)
+        self.vvn1_mongo_client.download_successful(self.guild_id, self.movie,final_dir_formatted)
         self.remove_torrents()
         return final_dir
 
@@ -77,8 +84,12 @@ class DownloadClient:
                 
         if first_sub_dir != '':
             shutil.rmtree(first_sub_dir)
+        
+        # TODO: can add a prop to the download thing for the dir name in this case
 
-        os.rename(outer_dir,new_outer_name)
+        if not os.path.exists(new_outer_name) and os.path.exists(outer_dir):
+            os.rename(outer_dir,new_outer_name)
+            self.vvn1_mongo_client.add_path_to_drq(self.guild_id,self.movie,new_outer_name)
 
         return new_outer_name
 
@@ -152,8 +163,8 @@ class DownloadClient:
             if not torrent:
                 raise Exception('No torrent found')
             cur_time = int(time.time()-start_time)
-            self.vvn1_mongo_client.update_download_progress(
-                self.guild_id, self.movie, round(torrent.percent_done*100, 1), cur_time)
+            self.vvn1_mongo_client.update_percent(
+                self.guild_id, self.vvn1_mongo_client.DOWNLOADING, self.movie,round(torrent.percent_done*100, 1), cur_time)
             time.sleep(1)
 
             if(torrent.percent_done == 1):
@@ -201,6 +212,33 @@ class DownloadClient:
             [self.torrent_client.torrent.remove(
                 torrent.id, False) for torrent in torrents]
 
+    def zip_movie(self,movie_path):
+        zip_name = self.movie['zipName']
+        zip_password = self.movie['zipPassword']
+
+        files = [f.path for f in os.scandir(movie_path) if f.is_file() and f.path.endswith('.zip')]
+        if len(files)>0:
+            #if we find the zip then we make sure it has the name we want
+            hd_zip_name = files[0]           
+            new_name = os.path.join(os.path.dirname(hd_zip_name)  , zip_name)
+            os.rename(hd_zip_name,new_name)
+            hd_zip_name = new_name
+        
+
+        else:
+
+            hd_zip_name = os.path.join(movie_path,zip_name)
+            zip_command = ['zip','-j','-r']
+            if zip_password!='':
+                zip_command.append('--password')
+                zip_command.append(zip_password)
+            
+            zip_command.append(zip_name)
+            zip_command.append('.')
+
+            subprocess.run(zip_command,stdout=subprocess.DEVNULL,cwd=movie_path)
+        
+        return hd_zip_name
 
 '''
 TODO:
